refactor: 변수명 표기법 적용
