+++ src/app/home-page/page.tsx
@@ -1,10 +1,9 @@
 "use client"
 
 import Page from "app/page"
-import HomeButton from "components/page/Home/home-button"
-import NavBar from "components/page/Home/navigator"
-import HomeTitle from "components/page/Home/title"
-
+import HomeTitle from "components/page/Home/HomeTitle"
+import HomeButton from "components/page/Home/HomeButton"
+import NavBar from "components/page/Home/NavBar"
 
 const HomePage = () => {
   return(
