+++ src/app/timetable-page/page.tsx
@@ -1,9 +1,9 @@
 "use client"
 
 import Page from 'app/page'
-import NavBar from 'components/page/Home/navigator'
-import TimeTable from 'components/page/Timetable'
-import Header from 'components/page/Timetable/Header'
+import NavBar from 'components/page/Home/NavBar'
+import TimeTable from 'components/page/TimeTable'
+import Header from 'components/page/TimeTable/Header'
 
 const TimetablePage = () => (
   <Page title="TimetablePage">
